2020/tests/dev/erikiva/aoc2020/Day24Test.kt
package dev.erikiva.aoc2020

import org.junit.jupiter.api.Test

import org.junit.jupiter.api.Assertions.*
import java.nio.file.Files
import java.nio.file.Paths

internal class Day24Test {

    @Test
    fun getSolutionTest() {
        val input = Files.readAllLines(Paths.get("tests/dev/erikiva/aoc2020/Day24TestInput.txt"))
        val day24 = Day24();
        assertEquals(10, day24.getSolution(input))
    }

    @Test
    fun getSolution() {
        val input = Files.readAllLines(Paths.get("tests/dev/erikiva/aoc2020/Day24Input.txt"))
        val day24 = Day24();
        assertEquals(277, day24.getSolution(input))
    }

    @Test
    fun getSolutionTestFinal() {
        val input = Files.readAllLines(Paths.get("tests/dev/erikiva/aoc2020/Day24TestInput.txt"))
        val day24 = Day24();
        assertEquals(2208, day24.getSolution2(input))
    }

    @Test
    fun getSolutionFinal() {
        val input = Files.readAllLines(Paths.get("tests/dev/erikiva/aoc2020/Day24Input.txt"))
        val day24 = Day24();
        assertEquals(3531, day24.getSolution2(input))
    }
}